Cambio de ID de long a int y nueva funcionalidad
Se ha cambiado el tipo de ID de `long` a `int` en varios controladores y DTOs, mejorando la eficiencia en la base de datos. Se han añadido nuevos controladores y comandos para gestionar `RegisterModalityWithStudents`, incluyendo la creación, actualización y obtención de registros de modalidad con estudiantes. Además, se han actualizado las configuraciones de las entidades en la base de datos para reflejar este cambio de tipo de ID. Se han creado nuevos DTOs y comandos para manejar la lógica de negocio relacionada con esta nueva funcionalidad, asegurando la consistencia en todas las referencias a IDs.

+++ Api/Controllers/ModalityController.cs
@@ -6,7 +6,7 @@ using Microsoft.AspNetCore.Mvc;
 
 namespace Api.Controllers
 {
-    public class ModalityController : GenericController<Modality, long, ModalityDto>
+    public class ModalityController : GenericController<Modality, int, ModalityDto>
     {
         public ModalityController(IMediator mediator) : base(mediator)
         {
