+++ Application/Shared/DTOs/RegisterModalityStudent/RegisterModalityStudentDto.cs
@@ -2,9 +2,9 @@ using System;
 
 namespace Application.Shared.DTOs.RegisterModalityStudent
 {
-    public class RegisterModalityStudentDto : BaseDto<long>
+    public class RegisterModalityStudentDto : BaseDto<int>
     {
-        public long IdRegisterModality { get; set; }
+        public int IdRegisterModality { get; set; }
         public int IdUser { get; set; }
         public new int? IdUserCreatedAt { get; set; }
     }
