+++ Domain/Entities/RegisterModalityState.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 
 namespace Domain.Entities
 {
-    public class RegisterModalityState : BaseEntity<long>
+    public class RegisterModalityState : BaseEntity<int>
     {
         public string Code { get; set; } = string.Empty;
         public string Name { get; set; } = string.Empty;
